Finished Mechanics and Map
I implemented shooting mechanics to make the ball shoot at a 45 degree
angle with respect to the horizontal based on change in mouse position
(click-and-drag). Collision detection works, but even continuous
collision detection does not fix involutarily having to shoot twice in
order to traverse the map. A beginning map is included and works well.
Camera mechanics could be better but currently follow the ball.